lab4/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dnspython==2.2.1
fonttools==4.29.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
maxminddb==2.2.0
networkx==2.6.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
texttable==1.6.4
urllib3==1.26.8
